requirements.txt
arrow==1.3.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
diff-match-patch==20241021
Django==5.1.3
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-import-export==4.3.3
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
ffmpeg==1.4
python-dateutil==2.9.0.post0
redis==5.2.0
rq==1.16.2
six==1.16.0
sqlparse==0.5.2
tablib==3.7.0
times==0.7
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
